♻️ [Refactoring]: figma-infoモジュールをディレクトリ構造に再編成
- src/config/figma-info.tsをsrc/config/figma-info/ディレクトリに移動
- runtime-config.tsのimportパスを更新
- figma-info.ts内のimportパスを修正（../../constantsに変更）
